% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.gbtbin.R
\name{add.gbtbin}
\alias{add.gbtbin}
\title{Take union of two gbtbin objects}
\usage{
\method{add}{gbtbin}(x1, x2)
}
\arguments{
\item{x1}{Object of class gbtbin}

\item{x2}{Object of class gbtbin}
}
\value{
Object of class gbtbin
}
\description{
Take union of two gbtbin objects. Equivalent to the R union function
}
\details{
Self explanatory...
}
\seealso{
\code{\link{lej}}
}

