% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Take union of two gbtbin objects}
\usage{
add(x1, x2)
}
\arguments{
\item{x1}{Object of class gbtbin}

\item{x2}{Object of class gbtbin}
}
\value{
Object of class gbtbin
}
\description{
Combine two gbtbin objects into a single bin. Equivalent to the R union
function
}
\seealso{
\code{\link{lej}}
}
