% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/binsFromMarkers.R
\name{binsFromMarkers}
\alias{binsFromMarkers}
\title{Create bin objects from taxonomic affiliation of markers}
\usage{
binsFromMarkers(x, marksource = "amphora2", taxon = "Order",
  out = "taxonBins", to.list = TRUE)
}
\arguments{
\item{x}{gbt object}

\item{marksource}{Name of marker source to use for the taxon assignment}

\item{taxon}{Taxonomic level to perform the binning (e.g. "Class", "Family")}

\item{out}{Prefix for output bin names}

\item{to.list}{Output bins to list of gbtbin objects (default TRUE)}
}
\value{
List of gbtbin objects
}
\description{
Given a gbt object with taxonomic markers, create a set of bin objects
corresponding to the affiliations, at a given taxonomic level (e.g. Order or
Family).
}
\seealso{
\code{\link{summaryLOB}}

\code{\link{importBins}}

\code{\link{exportBins}}

\code{\link{winnow}}
}

