% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/choosebin.R
\name{choosebin}
\alias{choosebin}
\title{Choose bin interactively from plot of gbt object}
\usage{
choosebin(x, ...)
}
\arguments{
\item{x}{Object of class gbt, used to generate the plot}

\item{slice}{The same slice parameter used to generate the plot}
}
\value{
Object of class gbtbin
}
\description{
Choose genome bin from GC-coverage or differential coverage plot of a
gbt object
}
\seealso{
\code{\link{plot.gbt}}
}

