% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choosebinPolygon.R
\name{choosebinPolygon}
\alias{choosebinPolygon}
\title{Choose bin from gbt object by defining polygon in coverage-GC or differential coverage plot}
\usage{
choosebinPolygon(x, ...)
}
\arguments{
\item{x}{Object of class gbt}

\item{slice}{Slice parameter for drawing the polygon}

\item{polygon}{Polygon that would define the bin}
}
\value{
Object of class gbtbin
}
\description{
Choose genome bin by specifying coordinates of a polygon from
GC-coverage or differential coverage plot of a gbt object
}
\seealso{
\code{\link{plot.gbt}}
}
