% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/countSingleFromTable.R
\name{countSingleFromTable}
\alias{countSingleFromTable}
\title{Tabulate objects and count how many singletons}
\usage{
countSingleFromTable(x)
}
\arguments{
\item{x}{Object of class data.frame or vector}
}
\value{
Numeric vector of length 1
}
\description{
Tabulate objects and count how many singletons
}
\keyword{internal}

