% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/exportBins.R
\name{exportBins}
\alias{exportBins}
\title{Export a list of gbtbin objects to external file}
\usage{
exportBins(x, useBinNames = TRUE, prefix = "gbtExport")
}
\arguments{
\item{x}{List of gbtbin objects (e.g. list(bin1,bin2,bin3))}

\item{useBinNames}{Use the bin names associated with the list (default TRUE)}

\item{prefix}{If useBinNames is FALSE, then use this prefix followed by a number for the exported file names}
}
\value{
External files corresponding to the individual bins
}
\description{
Takes a list of gbtbin objects, and exports files, each corresponding
to a single bin, with a list of contig names in that bin
}
\seealso{
\code{\link{write.gbtbin}}

\code{\link{importBins}}

\code{\link{summaryLOB}}
}

