% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastgFishing.R
\name{fastgFishing}
\alias{fastgFishing}
\title{Perform connectivity fishing with Fastg file}
\usage{
fastgFishing(x, bin, fastg.file, ...)
}
\arguments{
\item{x}{Object of class gbt, from which the bin object was defined}

\item{bin}{Object of class gbtbin, defined from the gbt object x}

\item{fastg.file}{Path to Fastg file for the metagenome assembly of x}

\item{save}{Logical - Save list of contigs of new bin to external file?}

\item{file}{File to save contig list}
}
\value{
Object of class gbtbin
}
\description{
Find contigs connected to existing gbtbin object, using connectivity from
an external Fastg file
}

