% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbt.R
\name{gbt}
\alias{gbt}
\title{Create new gbt object}
\usage{
gbt(covstats, mark, marksource, ssu, trna)
}
\arguments{
\item{covstats}{File(s) with coverage statistics of a metagenomic assembly;
output of pileup.sh in BBTools package (required). More than one file
can be specified with c().}

\item{mark}{Table of scaffolds with marker genes and taxonomic information
for each marker. E.g. AMPHORA2 or Phyla-AMPHORA marker sets and output
parsed by parse_phylotype_result.pl. If more than one marker set is
available, this can be a vector (optional)}

\item{marksource}{Name for the marker set given to parameter \code{mark}. If
more than one marker file is supplied, this can be a vector.}

\item{ssu}{Table of scaffolds with SSU rRNA genes, and taxonomic info
for each SSU. E.g. use get_ssu_for_genome_bin_tools.pl. (optional)}

\item{tnra}{Table of tRNA genes found in assembly. Can use the output from
tRNAscan-SE directly. (optional)}
}
\value{
Object of class gbt
}
\description{
Creates new gbt object from coverage data, taxonomic markers, and other
data
}
\details{
See documentation online https://github.com/kbseah/genome-bin-tools for
fuller instructions on generating the input files required.
}
\seealso{
\code{\link{gbtbin}} - Define bin from gbt object

\code{\link{choosebin}} - Choose bin interactively from plot

\code{\link{gbt_checkinput}} - Check input files before import
}
