% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbt_checkinput.R
\name{gbt_checkinput}
\alias{gbt_checkinput}
\title{Check input files for gbtools}
\usage{
gbt_checkinput(covstats, mark = NULL, ssu = NULL, trna = NULL,
  outdir = NULL, log = NULL, script.path = system.file("Perl",
  "input_validator.pl", package = "gbtools"))
}
\arguments{
\item{covstats}{File(s) with coverage statistics of a metagenomic assembly;
output of pileup.sh in BBTools package (required). More  than one file
can be specified with c().}

\item{mark}{Table of scaffolds with marker genes and taxonomic information
for each marker. E.g. AMPHORA2 or Phyla-AMPHORA marker sets and output
parsed by parse_phylotype_result.pl. More than one file can be
specified with c() (optional)}

\item{ssu}{Table of scaffolds with SSU rRNA genes, and taxonomic info
for each SSU. E.g. use get_ssu_for_genome_bin_tools.pl. (optional)}

\item{trna}{Table of tRNA genes found in assembly. Can use the output from
tRNAscan-SE directly. (optional)}

\item{outdir}{Folder to write fixed versions of the input file, where this
is possible.}

\item{log}{Path to write log file of the input validator script, which you
can read for more details (Default: current working folder)}

\item{script.path}{Path to input validator script. Default is to look in the
location where gbtools package is installed.}
}
\value{
data.frame of file names and errors found per file
}
\description{
Check input files for the \code{\link{gbt}} function before importing them.
This calls a Perl script, located in inst/Perl/input_validator.pl within the
R package source. The input validator script can also be called outside of
the R environment, e.g. if you wish to check the files before starting an R
session. Calling input_validator.pl without arguments will show usage help.
}
\details{
See documentation online https://github.com/kbseah/genome-bin-tools for
fuller instructions on formats of the input files required.
}
\seealso{
\code{\link{gbt}} - Import data and create gbt object for analysis
}
