% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbtbin.R
\name{gbtbin}
\alias{gbtbin}
\title{Create gbtbin from gbt object}
\usage{
gbtbin(shortlist, x, slice, taxon, points, save, file)
}
\arguments{
\item{shortlist}{List of contigs in the new bin (required)}

\item{x}{Object of class gbt}

\item{slice}{Numeric vector describing which coverage values were used to
draw the plot from which the bin was selected by choosebin. Use NA if
not calling this function from choosebin.}

\item{taxon}{For internal use}

\item{points}{Numeric vector describing points used to select the bin from
a plot by choosebin. Use NA if not calling this function from choosebin.}

\item{save}{Save list of contigs to external file? (logical, default FALSE)}

\item{file}{File name to save contigs, if save=TRUE.}
}
\value{
Object of class gbtbin
}
\description{
Creates object of class gbtbin from existing gbt object, representing a
genome bin defined from a metagenome assembly.
}
\details{
This function is called by \code{\link{choosebin}}, but can be called
directly by the user, given a shortlist of contigs in the bin.
}
\seealso{
\code{\link{gbt}}

\code{\link{choosebin}}
}
