% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generatePlotColors.R
\name{generatePlotColors}
\alias{generatePlotColors}
\title{Generates colors for marker gene phylotypes in plot}
\usage{
generatePlotColors(scaffold.stats, marker.list, taxon, consensus)
}
\arguments{
\item{scaffold.stats}{scaff data.frame from a gbt object}

\item{marker.list}{List of markers from markTab data.frame of gbt object}

\item{taxon}{Taxon names for which to generate colors}

\item{consensus}{If scaffold has more than one marker, take majority-rule
consensus taxonomic assignment? (logical)}
}
\value{
data.frame with color assignments for each scaffold
}
\description{
Generates colors for marker gene phylotypes in plot
}
\keyword{internal}

