% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generatePlotColors2.R
\name{generatePlotColors2}
\alias{generatePlotColors2}
\title{Generates colors for marker gene phylotypes in plot by cumulative weight}
\usage{
generatePlotColors2(scaffold.stats, marker.list, taxon, consensus, weightCutoff,
  markCustomPalette)
}
\arguments{
\item{scaffold.stats}{Scaff table from gbt object}

\item{marker.list}{markTab table from gbt object}

\item{taxon}{Taxonomic level to do coloring}

\item{consensus}{Logical - if taxon assignments conflict, take consensus?}

\item{weightCutoff}{Cutoff quantile for contig length*coverage weight (between 0 and 1)}

\item{markCustomPalette}{(Optional) data.frame with taxon names (column 1) and specific colors (column 2) to assign to them}
}
\value{
data.frame with color assignments for each scaffold
}
\description{
For each taxon, calculates the total contigLength*coverage, and assigns
colors for the top taxa which in total account for more than a specified
minimum weight.
}
\keyword{internal}

