% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getN50.R
\name{getN50}
\alias{getN50}
\title{Calculate N50 from contig lengths}
\usage{
getN50(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
N50 value - min contig length that contains half of total bases
}
\description{
Calculate N50 from contig lengths
}
\keyword{internal}

