% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/importBins.R
\name{importBins}
\alias{importBins}
\title{Create several bin objects simultaneously from a table}
\usage{
importBins(x, file)
}
\arguments{
\item{x}{Object of class gbt representing the parent metagenome}

\item{file}{Table of contig names and bin names, produced by the parse_bin_fasta_files.pl script}
}
\value{
Creates objects of class gbtbin as side effect
}
\description{
Given a table of contigs and the bins that they belong to, create gbtbin
objects with corresponding bins from the parent gbt object. The table is
generated by the accessory script parse_bin_fasta_files.pl
}
\seealso{
\code{\link{gbt}}

\code{\link{gbtbin}}
}

