% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lej.gbtbin.R
\name{lej.gbtbin}
\alias{lej.gbtbin}
\title{Take difference between two gbtbin objects}
\usage{
\method{lej}{gbtbin}(x1, x2)
}
\arguments{
\item{x1}{Object of class gbtbin}

\item{x2}{Object of class gbtbin}
}
\description{
Takes the reverse complement of two gbtbin objects. Equivalent to setdiff
in R, or left-exclusive-join in SQL. Non commutative!
}
\details{
Self explanatory...
}
\seealso{
\code{\link{add}}
}

