% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lej.R
\name{lej}
\alias{lej}
\title{Take difference between two gbtbin objects}
\usage{
lej(x1, x2)
}
\arguments{
\item{x1}{Object of class gbtbin}

\item{x2}{Object of class gbtbin}
}
\description{
Takes the reverse complement of two gbtbin objects. Equivalent to setdiff
in R, or left-exclusive-join in SQL. Non commutative!
}
\seealso{
\code{\link{add}}
}
