% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mergeOverlapBins.R
\name{mergeOverlapBins}
\alias{mergeOverlapBins}
\title{Merge two sets of bins}
\usage{
mergeOverlapBins(x, y, binNames.x = "", binNames.y = "", weight = TRUE,
  by = "y", mergeto = "x", threshold = "0.8", out = "mergedBin")
}
\arguments{
\item{x}{List of gbtbin objects (either list(bin1,bin2) or mget(vector_of_bin_names)}

\item{y}{Second list of gbtbin objects}

\item{binNames.x}{Vector of names for each bin in list x}

\item{binNames.y}{Vector of names for each bin in list y}

\item{weight}{Compare by number of bases shared (default TRUE), else by number of contigs}

\item{by}{Condition by fraction in y (default) or x}

\item{mergeto}{Merge bins in list x to list y ("x" - default) or vice-versa ("y")}

\item{threshold}{Threshold for fraction that must be shared before bins are merged (default 0.8)}

\item{out}{Name prefix of output bins (default "mergedBin")}
}
\value{
List of gbtbin objects containing the merged bins (if any)
}
\description{
Given two lists of gbtbin objects, compare them in terms of their shared
contigs and merge bins which exceed a certain overlap threshold.
}
\seealso{
\code{\link{tabOverlapBins}}

\code{\link{importBins}}

\code{\link{multiBinPlot}}
}

