% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeScaffMarker.R
\name{mergeScaffMarker}
\alias{mergeScaffMarker}
\title{Merge scaffolds table with marker table}
\usage{
mergeScaffMarker(scaffold.stats, marker.list, taxon, consensus = TRUE)
}
\arguments{
\item{scaffold.stats}{scaff data.frame from a gbt object}

\item{marker.list}{List of markers from markTab data.frame of gbt object}

\item{taxon}{Taxon names for which to generate colors}

\item{consensus}{If scaffold has more than one marker, take majority-rule
consensus taxonomic assignment? (logical)}
}
\value{
data.frame with marker data and scaffold statistics
}
\description{
Merge table of scaffold statistics (output from pileup.sh in BBMap package)
and table of marker statistics parsed by parse_phylotype_result.pl
}
\keyword{internal}
