% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiBinPlot.R
\name{multiBinPlot}
\alias{multiBinPlot}
\title{Plot gbt object and overlay multiple bins}
\usage{
multiBinPlot(x, bins, binNames = "", legend = FALSE, slice = 1,
  assemblyName = "", cutoff = 0, cols = NULL, log = "default",
  main = "default", xlab = "default", ylab = "default", xlim = NULL,
  ylim = NULL)
}
\arguments{
\item{x}{Object of class gbt}

\item{bins}{List of gbtbin objects}

\item{binNames}{Vector of names for the bins (only used to draw legend,
default none)}

\item{legend}{Draw legend? (logical, default FALSE)}

\item{slice}{For plotting coverage data, which sample to use?}

\item{assemblyName}{Name of the metagenome, for plot title}

\item{cutoff}{Minimum length to plot contigs (numeric, default 1000)}

\item{cols}{Custom palette for bins overlay (default Rainbow)}

\item{log}{Which axes should be logarithmic scale}

\item{main}{Custom label for plot title}

\item{xlab}{Custom label for x-axis label}

\item{ylab}{Custom label for y-axis label}

\item{xlim}{Custom x-axis range}

\item{ylim}{Custom y-axis range}
}
\value{
New graphics object and plot, or error message
}
\description{
Given a metagenome represented by a gbt object, and several bins derived
from that metagenome, this function allows quick plotting of the bins
simultaneously as overlays.
This is useful for checking the results of automatic binning software,
for example.
}
\details{
Note that plot overlays (markers, ssu, trna) and userAxis parameters are
not allowed for this type of plot, unlike the usual plot.gbt function,
because they would interfere with the bin overlay plotting.

The gbtbin objects to be plotted must be specified as a list, e.g.
list(bin1,bin2,bin3). If you use the c() function there will be an error.

You can specify a custom color palette for the bins, or the default of
rainbow palette will be used.
To include a legend with which color corresponds to which name, you can
specify the bin names with the binNames parameter. if this is not given,
then the bins are simply numbered sequentially in the legend, according
to the order they were specified in the bins= parameter.
}
\seealso{
\code{\link{plot.gbt}}
}
