% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pickBinPoints.R
\name{pickBinPoints}
\alias{pickBinPoints}
\title{Define polygon for genome bin on plot}
\usage{
pickBinPoints(num.points = 6, draw.polygon = TRUE)
}
\arguments{
\item{num.points}{Number of points in polygon (integer)}

\item{draw.polygon}{Draw polygon as overlay on plot (logical)}
}
\value{
numerical vector containing polygon vertices coordinates
}
\description{
Wrapper for locator() and polygon() to perform interactive binning on the
current plot. Returns the polygon vertices which can be used in
get.bin.stats()
}
\keyword{internal}

