% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.gbt.R
\name{plot.gbt}
\alias{plot.gbt}
\title{Plot object of class gbt}
\usage{
\method{plot}{gbt}(x, slice = 1, cutoff = 1000, taxon = "Class",
  assemblyName = "", marker = TRUE, marksource = "", gc = FALSE,
  ssu = FALSE, trna = FALSE, consensus = TRUE, legend = FALSE,
  textlabel = FALSE, col = "grey", log = "default", main = "default",
  xlab = "default", ylab = "default", ...)
}
\arguments{
\item{x}{Object of class gbt}

\item{slice}{For plotting coverage data, which sample to use? (see
Details section below)}

\item{cutoff}{Minimum length to plot contigs (numeric, default 1000)}

\item{taxon}{Taxonomic level for coloring the taxonomic markers, e.g.
"Class" or "Phylum". (default "Class")}

\item{assemblyName}{Name of the metagenome, for plot title}

\item{marker}{Color plot by taxon markers? (logical, default TRUE)}

\item{gc}{Color plot by GC% instead of taxon markers? Only used for
differential coverage plots, i.e. when two values are supplied
to the slice parameter. (logical, default FALSE)}

\item{ssu}{Draw markers for SSU genes? (logical, default FALSE)}

\item{trna}{Draw markers for tRNA genes? (logical, default FALSE)}

\item{consensus}{For contigs with more than one marker gene with conflict
in taxonomy, take majority rule consensus? (logical,
default TRUE)}

\item{legend}{Draw legend? (logical, default FALSE)}

\item{textlabel}{Label SSU markers with taxon? (logical, default FALSE)}

\item{col}{Color for points (default "grey")}

\item{log}{Which axes should be logarithmic scale}

\item{main}{Custom label for plot title}

\item{xlab}{Custom label for x-axis label}

\item{ylab}{Custom label for y-axis label}

\item{...}{See par for more plot options}
}
\value{
New graphics object and plot, or error message
}
\description{
Plot GC-coverage or differential coverage plots from a gbt object
}
\details{
The plot method for gbt objects can produce both GC-coverage and
differential coverage plots. A gbt object contains scaffold data and
annotations for a given metagenome, along with the coverage of each
scaffold for one or more samples. To plot GC-coverage plots, the sample
to use for the coverage data is specified by giving the number of the
sample to the slice parameter. To plot differential-coverage plots, the
numbers of the two samples for comparison are given to slice as a vector
e.g. c(1,2). The first sample is plotted as the x-axis and the second as
the y-axis. Supplying a vector with more than 2 elements, or a non-
numeric value will return an error.
Genome bins can be interactively chosen from a gbt plot with the
\code{\link{choosebin}} function; the same slice argument must be passed
to the choosebin function as the plot function, otherwise the results will
be meaningless!
}
\seealso{
\code{\link{gbt}}
}

