% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.gbtbin.R
\name{plot.gbtbin}
\alias{plot.gbtbin}
\title{Plot object of class gbtbin}
\usage{
\method{plot}{gbtbin}(x, slice = "default", ...)
}
\arguments{
\item{x}{Object of class gbt}

\item{slice}{For plotting coverage data, which sample to use? (see
Details section below)}

\item{...}{See par for more plot options}
}
\value{
New graphics window and a plot
}
\description{
Plot GC-coverage or differential coverage plots from a gbtbin object
}
\details{
See documentation on \code{\link{plot.gbt}} for more details on plotting
}
\seealso{
\code{\link{plot.gbt}}, \code{\link{points.gbtbin}}
}

