% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/points.gbtbin.R
\name{points.gbtbin}
\alias{points.gbtbin}
\title{Add points to plot of gbt or gbtbin object}
\usage{
\method{points}{gbtbin}(x, col = "black", slice = "default", cutoff = 0,
  pch = 20, symbolScale = "area", symbolScaleParam = 100, ...)
}
\arguments{
\item{x}{Object of class gbt}

\item{col}{Color for points (default "grey")}

\item{slice}{}

\item{cutoff}{Minimum length to plot contigs (numeric, default 1000)}

\item{symbolScale}{Scaling function to use for plot characters (options: "area", "length")}

\item{symbolScaleParam}{Scaling constant for plot chars (default: 100)}

\item{...}{See par for more plot options}
}
\description{
Add points to GC-coverage or differential coverage plots from a gbt object
}
\details{
See \code{\link{plot.gbt}} for more details on gbt plots.
}
\seealso{
\code{\link{plot.gbt}}, \code{\link{plot.gbtbin}}
}

