% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gbt.R
\name{print.gbt}
\alias{print.gbt}
\title{Print summary of gbt object}
\usage{
\method{print}{gbt}(x)
}
\arguments{
\item{x}{Object of class gbt}
}
\value{
data.frame summarizing contents of x
}
\description{
Print summary of gbt object
}
\seealso{
\code{\link{gbt}}, \code{\link{summary.gbt}}
}
