% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.gbtbin.R
\name{print.gbtbin}
\alias{print.gbtbin}
\title{Print summary of gbt objectbin}
\usage{
\method{print}{gbtbin}(x)
}
\arguments{
\item{x}{Object of class gbtbin}
}
\value{
data.frame summarizing contents of x
}
\description{
Print summary of gbt objectbin
}
\seealso{
\code{\link{gbtbin}}, \code{\link{summary.gbtbin}}
}

