% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOperation.R
\name{setOperation}
\alias{setOperation}
\title{Generic operation for merging and subsetting two gbtbin objects}
\usage{
setOperation(x1, x2, shortlist)
}
\arguments{
\item{x1}{Object of class gbtbin}

\item{x2}{Object of class gbtbin}

\item{shortlist}{Vector of contig IDs to make new bin}
}
\value{
Object of class gbtbin
}
\description{
Generic operation for merging and subsetting two gbtbin objects
}
\keyword{internal}
