% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.gbt.R
\name{summary.gbt}
\alias{summary.gbt}
\title{Print summary of gbt object}
\usage{
\method{summary}{gbt}(x)
}
\arguments{
\item{x}{Object of class gbt}
}
\value{
data.frame summarizing contents of x
}
\description{
Print summary of gbt object
}
\seealso{
\code{\link{gbt}}, \code{\link{print.gbt}}
}

