% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.gbtbin.R
\name{summary.gbtbin}
\alias{summary.gbtbin}
\title{Print summary of gbtbin object}
\usage{
\method{summary}{gbtbin}(x)
}
\arguments{
\item{x}{Object of class gbtbin}
}
\value{
data.frame summarizing contents of x
}
\description{
For a gbtbin object, summary() differs from print() in displaying more
information: if available, table of counts per marker gene (for each marker
source), and table of counts per tRNA type.
}
\seealso{
\code{\link{gbtbin}}, \code{\link{print.gbtbin}}
}

