% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabOverlapBins.R
\name{tabOverlapBins}
\alias{tabOverlapBins}
\title{Compare two sets of bins}
\usage{
tabOverlapBins(x, y, binNames.x = "", binNames.y = "", weight = TRUE,
  by = "x")
}
\arguments{
\item{x}{List of gbtbin objects (either list(bin1,bin2) or
mget(vector_of_bin_names)}

\item{y}{Second list of gbtbin objects}

\item{binNames.x}{Vector of names for each bin in list x}

\item{binNames.y}{Vector of names for each bin in list y}

\item{weight}{Compare by number of bases shared (default TRUE), else by
number of contigs}

\item{by}{Which list of bins to compute fraction of overlap: "x" (default),
"y", or "raw" (raw counts)}
}
\value{
Matrix reporting quantity of overlap between two sets of bins
}
\description{
Given two lists of gbtbin objects, compare them in terms of their shared
contigs and output a table.
}
\seealso{
\code{\link{mergeOverlapBins}}

\code{\link{importBins}}

\code{\link{multiBinPlot}}
}
