% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winnow.R
\name{winnow}
\alias{winnow}
\title{Subset a gbt or gbtbin object by GC, length, or coverage cutoffs}
\usage{
winnow(x, gc, len, covmin, covmax, slice, save, file)
}
\arguments{
\item{x}{Object of class gbt or gbtbin}

\item{gc}{Vector giving min and max GC values (default c(0,Inf))}

\item{len}{Vector giving min and max contig lengths (default c(0,Inf))}

\item{covmin}{Vector giving minimum coverage values per sample (default NA)}

\item{covmax}{Vector giving maximum coverage values per sample (default NA)}

\item{slice}{Vector of sample numbers for the coverage cutoffs (default NA)}

\item{save}{Save list of contigs to external file? (logical, defautl FALSE)}

\item{file}{File name for export of contig list.}
}
\value{
Object of class gbtbin
}
\description{
Subset a gbt or gbtbin object by GC, length, or coverage cutoffs
}
\seealso{
\code{\link{winnowMark}}, \code{\link{gbt}}
}
