% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winnowMark.R
\name{winnowMark}
\alias{winnowMark}
\title{Subset a gbt or gbtbin object by marker gene taxonomy}
\usage{
winnowMark(x, marksource = NA, param = "Class",
  value = "Gammaproteobacteria", save = FALSE, file)
}
\arguments{
\item{x}{Object of class gbt or gbtbin}

\item{marksource}{Which marker set to use (default -- use the first one)}

\item{param}{Taxonomic level to make subset (default "Class")}

\item{value}{Value of the taxon to make subset (default "Gammaproteobacteria")}

\item{save}{Logical: save shortlist of contigs to external file? (default: FALSE)}

\item{file}{File to save shortlist of contigs}
}
\value{
Object of class gbtbin
}
\description{
Subset a gbt or gbtbin object by marker gene taxonomy
}
\seealso{
\code{\link{winnow}}, \code{\link{gbt}}
}
