% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.gbt.R
\name{write.gbt}
\alias{write.gbt}
\title{Export contig list from gbt or gbtbin object}
\usage{
write.gbt(x, file = "gbtools_export.list")
}
\arguments{
\item{x}{Object of class gbt or gbtbin}

\item{file}{Path to file for writing data (default "gbtools_export")}
}
\value{
External file with list of contigs separated by newlines
}
\description{
Export contig list from gbt or gbtbin object
}
\seealso{
\code{\link{gbt}}, \code{\link{gbtbin}}
}
