% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cexScaling.R
\name{cexScaling}
\alias{cexScaling}
\title{Function for different types of plot character scaling}
\usage{
cexScaling(x, type, const)
}
\description{
Function for different types of plot character scaling
}
\seealso{
\code{\link{plot.gbt}}
}
