% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gbtools.R
\docType{package}
\name{gbtools}
\alias{gbtools}
\alias{gbtools-package}
\title{Interactive visualization of metagenome assemblies in R}
\description{
gbtools lets you visualize metagenome assemblies by GC-coverage or
differential coverage plots, and interactively bin them.
}
\details{
See website for more details.
}
\author{
Brandon Seah, \email{kbseah@mpi-bremen.de}
}
\references{
\url{https://github.com/kbseah/genome-bin-tools}
}

