% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gbtbin.R
\name{plot.gbtbin}
\alias{plot.gbtbin}
\title{Plot object of class gbtbin}
\usage{
\method{plot}{gbtbin}(x, slice = 1, cutoff = 1000, taxon = "Class",
  taxonLevel = "", highlightTaxon = "", assemblyName = "",
  marker = TRUE, marksource = "", markCutoff = 0.99,
  markCustomPalette = 0, gc = FALSE, ssu = FALSE, trna = FALSE,
  consensus = TRUE, legend = FALSE, textlabel = FALSE, userAxis = NA,
  col = "grey", log = "default", main = "default", xlab = "default",
  ylab = "default", symbolScale = "area", symbolScaleParam = 100, ...)
}
\arguments{
\item{x}{Object of class gbt or gbtbin}

\item{slice}{For plotting coverage data, which sample to use? (see Details
section below)}

\item{cutoff}{Minimum length to plot contigs (numeric, default 1000)}

\item{taxonLevel}{Taxonomic level for coloring the taxonomic markers, e.g.
"Class" or "Phylum". (default "Class")}

\item{highlightTaxon}{Color markers affiliated to specified taxon only
(default: "")}

\item{assemblyName}{Name of the metagenome, for plot title}

\item{marker}{Color plot by taxon markers? (logical, default TRUE)}

\item{marksource}{Specify which marker set to plot (default: first supplied)}

\item{markCutoff}{Length x coverage weight cutoff for colored markers
(default: 0.99)}

\item{markCustomPalette}{(Optional) Assign colors to use for markers
belonging to specific taxa. This must be a data frame with taxon names
in column 1, and color names (English or hex codes) in column 2. Taxa
that are not represented in the gbt object will be ignored. Taxa not
in this table will be colored grey. If custom colors are specified,
then the markCutoff parameter will be ignored. (default: 0)}

\item{gc}{Color plot by GC% instead of taxon markers? Only used for
differential coverage plots, i.e. when two values are supplied to the
slice parameter. (logical, default FALSE)}

\item{ssu}{Draw markers for SSU genes? (logical, default FALSE)}

\item{trna}{Draw markers for tRNA genes? (logical, default FALSE)}

\item{consensus}{For contigs with more than one marker gene with conflict
in taxonomy, take majority rule consensus? (logical, default TRUE)}

\item{legend}{Draw legend? (logical, default FALSE)}

\item{textlabel}{Label SSU markers with taxon? (logical, default FALSE)}

\item{userAxis}{Use user-custom values for axis. See Details}

\item{col}{Color for points (default "grey")}

\item{log}{Which axes should be logarithmic scale}

\item{main}{Custom label for plot title}

\item{xlab}{Custom label for x-axis label}

\item{ylab}{Custom label for y-axis label}

\item{symbolScale}{Scaling function to use for plot characters (options:
"area", "length")}

\item{symbolScaleParam}{Scaling constant for plot chars (default: 100)}

\item{...}{See par for more plot options}
}
\value{
New graphics object and plot, or error message
}
\description{
Plot GC-coverage or differential coverage plots from a gbt object
}
\details{
The plot method for gbt objects can produce both GC-coverage and
differential coverage plots. A gbt object contains scaffold data and
annotations for a given metagenome, along with the coverage of each
scaffold for one or more samples. To plot GC-coverage plots, the sample
to use for the coverage data is specified by giving the number of the
sample to the slice parameter. To plot differential-coverage plots, the
numbers of the two samples for comparison are given to slice as a vector
e.g. c(1,2). The first sample is plotted as the x-axis and the second as
the y-axis. Supplying a vector with more than 2 elements, or a non-
numeric value will return an error.
Genome bins can be interactively chosen from a gbt plot with the
\code{\link{choosebin}} function; the same slice argument must be passed
to the choosebin function as the plot function, otherwise the results will
be meaningless!
If you wish to plot by user-supplied custom values (added to gbt object by
the userAdd() function), indicate which ones to use for x- or y- axis in
the userAxis= parameter. For example, to plot user-custom data set 1 in X-
and user-custom data set 2 in Y-axis, specify userAxis=c(1,2). To plot vs.
GC or coverage values, supply "gc" or "cov" to userAxis, e.g. GC as X-axis
and user-value set 2 as Y-axis: userAxis=c("gc",2). If "cov" specified, it
will take the coverage set that is given by the slice= parameter (default
is the first set of coverage data).
}
\seealso{
\code{\link{gbt}}
}
